hoist `export` out of `ChainStore`
