+++ src/chain/store/chain_store.rs
@@ -6,7 +6,6 @@ use std::{path::Path, sync::Arc};
 use crate::beacon::{BeaconEntry, IGNORE_DRAND_VAR};
 use crate::blocks::{BlockHeader, Tipset, TipsetKeys, TxMeta};
 use crate::interpreter::BlockMessages;
-use crate::ipld::stream_chain;
 use crate::libp2p_bitswap::{BitswapStoreRead, BitswapStoreReadWrite};
 use crate::message::{ChainMessage, Message as MessageTrait, SignedMessage};
 use crate::networks::ChainConfig;
@@ -15,24 +14,15 @@ use crate::shim::{
     address::Address, econ::TokenAmount, executor::Receipt, message::Message,
     state_tree::StateTree, version::NetworkVersion,
 };
-use crate::utils::{
-    db::{
-        file_backed_obj::{ChainMeta, FileBacked},
-        BlockstoreExt, CborStoreExt,
-    },
-    io::{AsyncWriterWithChecksum, Checksum},
+use crate::utils::db::{
+    file_backed_obj::{ChainMeta, FileBacked},
+    BlockstoreExt, CborStoreExt,
 };
 use ahash::{HashMap, HashMapExt, HashSet};
-use anyhow::{Context, Result};
-use async_compression::futures::write::ZstdEncoder;
+use anyhow::Result;
 use cid::Cid;
-use digest::Digest;
-use futures::{io::BufWriter, AsyncWrite};
-use futures_util::future::Either;
-use futures_util::AsyncWriteExt;
 use fvm_ipld_amt::Amtv0 as Amt;
 use fvm_ipld_blockstore::Blockstore;
-use fvm_ipld_car::CarHeader;
 use fvm_ipld_encoding::CborStore;
 use parking_lot::Mutex;
 use serde::{de::DeserializeOwned, Serialize};
@@ -385,53 +375,6 @@ where
     pub fn messages_for_tipset(&self, ts: &Tipset) -> Result<Vec<ChainMessage>, Error> {
         let bmsgs = self.block_msgs_for_tipset(ts)?;
         Ok(bmsgs.into_iter().flat_map(|bm| bm.messages).collect())
-    }
-
-    pub async fn export<W, D>(
-        &self,
-        tipset: &Tipset,
-        lookup_depth: ChainEpochDelta,
-        writer: W,
-        compressed: bool,
-        skip_checksum: bool,
-    ) -> Result<Option<digest::Output<D>>, Error>
-    where
-        DB: Send + Sync,
-        D: Digest + Send + 'static,
-        W: AsyncWrite + Send + Unpin + 'static,
-    {
-        use futures::StreamExt;
-        let writer = AsyncWriterWithChecksum::<D, _>::new(BufWriter::new(writer), !skip_checksum);
-        let mut writer = if compressed {
-            Either::Left(ZstdEncoder::new(writer))
-        } else {
-            Either::Right(writer)
-        };
-
-        let stateroot_lookup_limit = tipset.epoch() - lookup_depth;
-
-        let mut stream = stream_chain(
-            self.db.clone(),
-            tipset.clone().chain(self.db.clone()),
-            stateroot_lookup_limit,
-        )
-        .map(|result| result.unwrap()); // FIXME: use a sink that supports TryStream.
-        let header = CarHeader::from(tipset.key().cids().to_vec());
-        header
-            .write_stream_async(&mut writer, &mut stream)
-            .await
-            .map_err(|e| Error::Other(format!("Failed to write blocks in export: {e}")))?;
-
-        writer.flush().await.context("failed to flush")?;
-        writer.close().await.context("failed to close")?;
-
-        let digest = match &mut writer {
-            Either::Left(left) => left.get_mut().finalize().await,
-            Either::Right(right) => right.finalize().await,
-        }
-        .map_err(|e| Error::Other(e.to_string()))?;
-
-        Ok(digest)
     }
 
     /// Get the [`TipsetKeys`] for a given epoch. The returned key will never be null.
